Fix integer types detection in case of hiding by typedefs
Current implementation of formatting relies on checking whether the
requested variable is integer or not. However, the function performing
such analysis ignores that type may be an alias produced by typedef.
It leads to incorrect formatting of integers having non-primitive types.

So, this commit forces the existing checker to use canonical types
instead of raw ones.